+++ app/_features/home/create-room/use-create-room.ts
@@ -1,7 +1,7 @@
 import { useNavigate } from '@/_shared/hooks/use-navigate';
 import { Mixpanel } from '@/_shared/components/analytics/mixpanel';
 import { room } from '@/_shared/utils/sdk';
-import { createHostCookieAction } from '@/_features/home/create-room/create-room-action';
+import { createHostCookieAction } from '@/_features/home/create-room/create-room-server-action';
 
 export const useCreateRoom = () => {
   const { navigateTo } = useNavigate();
